requirements.txt
Flask==2.3.2
requests==2.31.0
python-dateutil==2.8.2
plotly==5.15.0